Refactor DeviceData into UidConverter Device data and lookup only ever used inside the UidConverter
